requirements.txt
setuptools==30.1.0
python_vlc==1.1.2
